# api-spec.yml
swagger: "2.0"
info:
  description: Basic example of putting together an OpenAPI spec (Swagger v2 format).
  version: "1"
  title: "Demo OpenAPI Spec"
consumes:
  - application/json
produces:
  - application/json
...
host: localhost:4200
basePath: /api/v1
schemes:
  - https
  - http
paths:
  /users:
    post:
      tags:
        - users
      operationId: add-user
      summary: Add a new User
      parameters:
        - name: user
          in: body
          required: true
          description: New User Data
          schema:
            $ref: '#/definitions/user'
      responses:
        '204':
          description: User created.
        '400':
          description: Invalid data.
        '500':
          description: Server Error.
    get:
      tags:
        - users
      operationId: get-users
      summary: List of all Users
      responses:
        '200':
          description: Returns a JSON Array of User Objects
          schema:
            $ref: "#/definitions/userList"
    put:
      tags:
        - users
      operationId: put-users
      summary: Modify all Users
      responses:
        '400':
          description: Feature not enabled
    delete:
      tags:
        - users
      operationId: delete-users
      summary: Delete all Users
      responses:
        '400':
          description: Feature not enabled
  /users/{userID}:
    get:
      tags:
        - users
      operationId: get-user
      summary: Get a Specific User by ID
      parameters:
        - name: userID
          in: path
          description: User Specific ID
          required: true
          type: integer
      responses:
        '200':
          description: Returns a User Object
          schema:
            $ref: "#/definitions/user"
        '400':
          description: Invalid ID supplied
        '404':
          description: User not found
    put:
      tags:
        - users
      operationId: put-user
      summary: Update a Specific User by ID
      parameters:
        - name: userID
          in: path
          description: User Specific ID
          required: true
          type: integer
        - name: user
          in: body
          required: true
          description: New User Data
          schema:
            $ref: '#/definitions/user'
      responses:
        '204':
          description: User Modified.
        '400':
          description: Invalid data.
        '500':
          description: Server Error.
    delete:
      tags:
        - users
      operationId: delete-user
      summary: Delete a Specific User by ID
      parameters:
        - name: userID
          in: path
          description: User Specific ID
          required: true
          type: integer
      responses:
        '204':
          description: User Deleted.
...
definitions:
  user:
    type: object
    properties:
      id:
        type: integer
        description: User ID
      username:
        type: string
        description: User Name
      password:
        type: string
        description: User Password
      apikey:
        type: string
        format: uuid
        description: A user specific API Key
      isadmin:
        type: boolean
        description: Is the user an Admin
  userList:
    type: object
    properties:
      total:
        type: integer
      users:
        type: array
        items:
          $ref: "#/definitions/user"
        description: Array of Widgets
...
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: header
    name: X-API-Key
  OAuth2:
    type: oauth2
    flow: accessCode
    authorizationUrl: https://example.com/oauth/authorize
    tokenUrl: https://example.com/oauth/token
    scopes:
      view_users: view users
      admin_users: admin all users
      view_widgets: view widgets
      admin_widgets: admin widgets
      view_orders: view orders
      admin_orders: admin orders