# api/express/routes/api.js
// Import dependencies
const mongoose = require('mongoose');
const express = require('express');
const router = express.Router();

// MongoDB URL from the docker-compose file
const dbHost = 'mongodb://database/testing-data';

// Connect to mongodb
mongoose.connect(dbHost);

// create mongoose schema
const userSchema = new mongoose.Schema({
	id: Number,
	username: String,
	password: String,
	apikey: String,
	isadmin: Boolean
});

// create mongoose model
const User = mongoose.model('User', userSchema);

/* GET api listing. */
router.get('/', (req, res) => {
	res.send('api works');
});

/* GET all users. */
router.get('/users', (req, res) => {
	User.find({}, (err, users) => {
	if (err) res.status(500).send(error)

res.status(200).json(users);
});
});

/* GET one users. */
router.get('/users/:id', (req, res) => {
	User.findById(req.param.id, (err, users) => {
	if (err) res.status(500).send(error)

res.status(200).json(users);
});
});

/* Create a user. */
router.post('/users', (req, res) => {
	let user = new User({
		id: req.body.id,
		username: req.body.username,
		password: req.body.password,
		apikey: req.body.apikey,
		isadmin: req.body.isadmin
	});

user.save(error => {
	if (error) res.status(500).send(error);

res.status(201).json({
	message: 'User created successfully'
});
});
});

module.exports = router;